Move S3 classes to s3 project
