+++ core/build.sbt
@@ -4,7 +4,6 @@ dependencyOverrides += "org.scala-lang" % "scala-reflect" % "2.11.5"
 
 libraryDependencies ++= Seq(
   sprayJson,
-  awsJavaSdk,
   commonsIO,
   ai2Common,
   allenAiTestkit % "test",
