+++ core/src/main/scala/org/allenai/pipeline/ArtifactFactory.scala
@@ -3,8 +3,6 @@ package org.allenai.pipeline
 import java.io.File
 import java.net.URI
 
-import com.amazonaws.auth.BasicAWSCredentials
-
 import scala.reflect.ClassTag
 
 /** Creates an Artifact from a String
@@ -117,66 +115,16 @@ object UrlToArtifact {
     }
   }
 
-  // Create a FlatArtifact or StructuredArtifact from an absolute s3:// URL
-  def absoluteS3(credentials: => BasicAWSCredentials) = new UrlToArtifact {
-    def urlToArtifact[A <: Artifact: ClassTag]: PartialFunction[URI, A] = {
-      val c = implicitly[ClassTag[A]].runtimeClass.asInstanceOf[Class[A]]
-      val fn: PartialFunction[URI, A] = {
-        case url if c.isAssignableFrom(classOf[S3FlatArtifact])
-          && List("s3", "s3n").contains(url.getScheme) =>
-          val bucket = url.getHost
-          val path = url.getPath.dropWhile(_ == '/')
-          new S3FlatArtifact(path, S3Config(bucket, credentials)).asInstanceOf[A]
-        case url if c.isAssignableFrom(classOf[S3ZipArtifact])
-          && List("s3", "s3n").contains(url.getScheme) =>
-          val bucket = url.getHost
-          val path = url.getPath.dropWhile(_ == '/')
-          new S3ZipArtifact(path, S3Config(bucket, credentials)).asInstanceOf[A]
-      }
-      fn
-    }
-  }
-
-  // Create a FlatArtifact or StructuredArtifact from a path relative to the input s3:// URL
-  def relativeS3(
-    cfg: => S3Config,
-    rootPath: String
-  ) = new UrlToArtifact {
-    val cleanRoot = rootPath.dropWhile(_ == '/').reverse.dropWhile(_ == '/').reverse
-    def urlToArtifact[A <: Artifact: ClassTag]: PartialFunction[URI, A] = {
-      val c = implicitly[ClassTag[A]].runtimeClass.asInstanceOf[Class[A]]
-      val fn: PartialFunction[URI, A] = {
-        case url if c.isAssignableFrom(classOf[S3FlatArtifact])
-          && url.getScheme == null =>
-          val absPath = s"$cleanRoot/${url.getPath}"
-          new S3FlatArtifact(absPath, cfg).asInstanceOf[A]
-        case url if c.isAssignableFrom(classOf[S3ZipArtifact])
-          && url.getScheme == null =>
-          val absPath = s"$cleanRoot/${url.getPath}"
-          new S3ZipArtifact(absPath, cfg).asInstanceOf[A]
-      }
-      fn
-    }
-  }
-
-  // Create a FlatArtifact or StructuredArtfact from an input file:// or s3:// URL
-  def absoluteUrl(credentials: => BasicAWSCredentials = S3Config.environmentCredentials): UrlToArtifact =
-    UrlToArtifact.chain(absoluteFile, absoluteS3(credentials))
-
   // Create a FlatArtifact or StructuredArtifact from a path relative to the input file:// or s3:// URL
   def relativeToUrl[A <: Artifact: ClassTag](
-    rootUrl: URI,
-    credentials: => BasicAWSCredentials = S3Config.environmentCredentials
+    rootUrl: URI
   ): UrlToArtifact = {
     rootUrl match {
       case url if url.getScheme == "file" || url.getScheme == null =>
         val rootDir = new File(url.getPath)
-        UrlToArtifact.chain(relativeFile(rootDir), absoluteFile)
-      case url if url.getScheme == "s3" || url.getScheme == "s3n" =>
-        val cfg = S3Config(url.getHost, credentials)
-        val rootPath = url.getPath
-        UrlToArtifact.chain(relativeS3(cfg, rootPath), absoluteS3(credentials))
+        chain(relativeFile(rootDir), absoluteFile)
     }
   }
+
 }
